requirements.txt
python3-customtkinter>=4.1.0
os>=3.11
random>=3.11
json>=3.11
tkinter>=3.11